requirements.txt
streamlit==1.31.0
pillow==10.2.0
pytesseract==0.3.10
numpy==1.26.3